requirements.txt
pypdf
langchain
chromadb # Vector storage
pytest
boto3
langchain-community
python-docx
pandas
umap
plotly
sklearn
numpy
